requirements.txt

# File analysis
python-magic==0.4.27
pyelftools==0.32

# Static analysis
yara-python==4.5.2
pycryptodome==3.23.0

# Reporting
markdown2==2.5.3
jinja2==3.1.6
